% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dante_corpora.R
\name{get_corpus}
\alias{get_corpus}
\title{Retornar um corpus}
\usage{
get_corpus(corpus, layer)
}
\arguments{
\item{corpus}{O nome do corpus que se deseja retornar. Deve ser uma string, como "dantestocks" ou "danteshots".}

\item{layer}{A camada de anotação que se deseja retornar. Deve ser uma string, para o caso "dantestocks" temos as opções "base", "emotion" e "pos", já para a opção "dantestocks" temos as opções "base", "stance" e "pos".}
}
\value{
Um data.frame contendo os dados do arquivo do corpus escolhido.
}
\description{
Retorna qualquer um dos corpora do projeto DANTE do POeTiSA no formato data.frame do R.
}
\examples{
# Exemplo: Retornar o arquivo "pos" do corpus "dantestocks"
df_pos_stocks <- get_corpus("dantestocks", "pos")
print(df_pos_stocks)
}
